+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add LSP completion for `buf.gen.yaml`, `buf.yaml`, and `buf.policy.yaml` files.
 - Add error for when a dependency is added to `buf.yaml` and is missing from `buf.lock`.
+- Add `buf beta registry cargo` subcommand that acts as a Cargo `cargo:token-from-stdout` credential provider for BSR-hosted Cargo registries.
 
 ## [v1.69.0] - 2026-04-29